While Caro en el borrado

+++ app/Console/Commands/BorrarDataSeguimiento.php
@@ -43,19 +43,33 @@ class BorrarDataSeguimiento extends Command
     {
         $cuenta_id = $this->argument('cuenta_id');
         $this->info('Inicio ejecución: '.Carbon::now());
-        $data_count = 0;
-        $etapas = DB::table('etapa')
+        $amount = 10000; // cantidad de registros a operar
+        $etapas_left = true; // partimos asumiendo que se encuentran registros
+        $count = 0; // cuenta cuantas veces iteramos en el while
+        $total_data_count = 0; // total de dato_seguimiento borrados
+        while ($etapas_left) {
+            $offset = $count * $amount; // parte en cero, aumenta de 10.000 en 10.000
+            $etapas = DB::table('etapa')
                     ->select('etapa.id')
                     ->leftJoin('tramite', 'etapa.tramite_id', '=', 'tramite.id')
                     ->leftJoin('proceso', 'tramite.proceso_id', '=', 'proceso.id')
                     ->where('proceso.cuenta_id',$cuenta_id)
-                    ->take(10000)
+                    ->skip($offset) // skip es como el offset pero cuando va junto con take()
+                    ->take($amount)
                     ->orderBy('etapa.id', 'DESC')
                     ->get()->toArray();
-        $etapas = json_decode(json_encode($etapas), true);
-        $data_count = DB::table('dato_seguimiento')->whereIn('etapa_id', $etapas)->count();
-        $eliminados = DB::table('dato_seguimiento')->whereIn('etapa_id', $etapas)->delete();
-        $this->info("Registros eliminados de dato_seguimiento: ".$data_count);
+           if (count($etapas) > 0){
+                $etapas = json_decode(json_encode($etapas), true);
+                $data_count = DB::table('dato_seguimiento')->whereIn('etapa_id', $etapas)->count();
+                $eliminados = DB::table('dato_seguimiento')->whereIn('etapa_id', $etapas)->delete();
+                $this->info("Registros eliminados de dato_seguimiento: ".$data_count);
+                $total_data_count += $data_count;
+                $count++;
+            } else {
+                $etapas_left = false; // no encontramos más registros, nos permite salir del while
+            }
+        }
+        $this->info("Total registros eliminados de dato_seguimiento: ".$total_data_count);
         $this->info('Fin ejecución: '.Carbon::now());
     }
 }
